Renamed Pieces package to Entity. Implemented makeMove method in GameManager, modified removePiece method in Board class to return the removed piece.
