+++ app/src/main/java/Entities/King.java
@@ -1,4 +1,4 @@
-package Pieces;
+package Entities;
 
 public class King extends Piece {
