+++ app/src/main/java/Entities/Knight.java
@@ -1,4 +1,4 @@
-package Pieces;
+package Entities;
 
 public class Knight extends Piece {
