+++ app/src/main/java/Entities/Piece.java
@@ -1,4 +1,4 @@
-package Pieces;
+package Entities;
 
 public class Piece {
