+++ app/src/main/java/Entities/Rook.java
@@ -1,4 +1,4 @@
-package Pieces;
+package Entities;
 
 public class Rook extends Piece {
