+++ app/src/main/java/UseCases/NormalGameBuilder.java
@@ -1,12 +1,12 @@
 package UseCases;
 
-import Pieces.Bishop;
-import Pieces.Board;
-import Pieces.King;
-import Pieces.Knight;
-import Pieces.Pawn;
-import Pieces.Queen;
-import Pieces.Rook;
+import Entities.Bishop;
+import Entities.Board;
+import Entities.King;
+import Entities.Knight;
+import Entities.Pawn;
+import Entities.Queen;
+import Entities.Rook;
 import Database.User;
 
 class NormalGameBuilder implements GameBuilder {
